pyproject.toml
[tool.poetry]
name = "imagesecrets-api"
version = "0.1.0"
description = "A tool for encoding and decoding messages into/from images."
authors = ["Lukas Kucera <lukas.kucera.g@protonmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://gitlab.com/Lkms19/imagesecrets-api"

[tool.black]
line-length = 88
include = '^.+\.py$'

[tool.isort]
profile = "black"
src_paths = ["image_secrets", "tests"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.2.0"
fastapi = "^0.65.0"
numpy = "^1.20.2"
uvicorn = "^0.13.4"
python-multipart = "^0.0.5"
requests = "^2.25.1"
aiofiles = "^0.6.0"
pydantic = {extras = ["email"], version = "^1.8.2"}
SQLAlchemy = "^1.4.15"
asyncpg = "^0.22.0"
python-jose = "^3.2.0"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pytest-mock = "^3.6.1"
python-dotenv = "^0.17.1"
httpx = "^0.18.1"
pytest-asyncio = "^0.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
